Uploaded json->gif converting scripts
